Assembly routines in libc modified for compatibility with new /bin/as. Some gcc warnings fixed.

+++ include/strings.h
@@ -34,7 +34,7 @@
 
 int      bcmp(const void *, const void *, size_t);
 void     bcopy(const void *, void *, size_t);
-void     bzero(void *, size_t);
+void     bzero(void *, unsigned long);
 int      ffs(int);
 char    *index(const char *, int);
 char    *rindex(const char *, int);
